Factor into vb2cs specific code
