+++ ICSharpCode.CodeConverter/CSharp/VBToCSConversion.cs
@@ -38,13 +38,15 @@ namespace ICSharpCode.CodeConverter.CSharp
 
         public string RootNamespace { get; set; }
 
-        public async Task Initialize(Project project)
+        public async Task<Project> InitializeSource(Project project)
         {
+            project = await project.WithRenamedMergedMyNamespace();
             _sourceVbProject = project;
             var cSharpCompilationOptions = CSharpCompiler.CreateCompilationOptions();
             _convertedCsProject = project.ToProjectFromAnyOptions(cSharpCompilationOptions, DoNotAllowImplicitDefault);
             _csharpReferenceProject = project.CreateReferenceOnlyProjectFromAnyOptionsAsync(cSharpCompilationOptions);
             _csharpViewOfVbSymbols = (CSharpCompilation) await _csharpReferenceProject.GetCompilationAsync();
+            return project;
         }
 
         public async Task<SyntaxNode> SingleFirstPass(Document document)
@@ -95,9 +97,11 @@ namespace ICSharpCode.CodeConverter.CSharp
                    xml.Substring(defineConstantsEnd);
         }
 
-        public Project GetConvertedProject()
+        public async Task<(Project project, List<(string Path, DocumentId DocId, string[] Errors)> firstPassDocIds)>
+            GetConvertedProject((string Path, SyntaxNode Node, string[] Errors)[] firstPassResults)
         {
-            return _convertedCsProject;
+            var (project, docIds) = _convertedCsProject.WithDocuments(firstPassResults);
+            return (await project.RenameMergedMyNamespace(), docIds);
         }
 
         public string TargetLanguage { get; } = LanguageNames.CSharp;
